requirements.txt
# Runtime dependencies
websockets>=12.0
requests>=2.31.0
PyYAML>=6.0

# Development / testing
pytest>=8.0
pytest-asyncio>=0.23
